Kleine cleanups
