+++ lib/controller/LoginController.php
@@ -3,18 +3,16 @@
 namespace CsrDelft\controller;
 
 use CsrDelft\common\Annotation\Auth;
-use CsrDelft\repository\ProfielRepository;
 use CsrDelft\repository\security\RememberLoginRepository;
 use CsrDelft\service\security\LoginService;
 use CsrDelft\service\security\SuService;
 use CsrDelft\view\login\LoginForm;
-use Doctrine\ORM\OptimisticLockException;
-use Doctrine\ORM\ORMException;
-use Exception;
-use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
+use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
 use Symfony\Component\Security\Http\Util\TargetPathTrait;
 
 /**
@@ -26,6 +24,7 @@ use Symfony\Component\Security\Http\Util\TargetPathTrait;
  */
 class LoginController extends AbstractController {
 	use TargetPathTrait;
+
 	/**
 	 * @var LoginService
 	 */
@@ -51,7 +50,7 @@ class LoginController extends AbstractController {
 	 * @Route("/login", methods={"GET"})
 	 * @Auth(P_PUBLIC)
 	 */
-	public function loginForm (Request $request) {
+	public function loginForm(Request $request, UrlGeneratorInterface $urlGenerator, CsrfTokenManagerInterface $csrfTokenManager) {
 		if ($this->getUser()) {
 			return $this->redirectToRoute('default');
 		}
@@ -61,7 +60,9 @@ class LoginController extends AbstractController {
 			$this->saveTargetPath($request->getSession(), 'main', $targetPath);
 		}
 
-		$response = new Response(view('layout-extern.login', ['loginForm' => new LoginForm()]));
+		$response = new Response(view('layout-extern.login', [
+			'loginForm' => new LoginForm($urlGenerator, $csrfTokenManager->getToken('authenticate'), $this->loginService->getError())
+		]));
 
 		// Als er geredirect wordt, stuur dan een forbidden status
 		if ($targetPath) {
@@ -71,46 +72,38 @@ class LoginController extends AbstractController {
 		return $response;
 	}
 
-//	/**
-//	 * @return RedirectResponse
-//	 * @throws ORMException
-//	 * @throws OptimisticLockException
-//	 * @throws Exception
-//	 * @Route("/login", methods={"POST"})
-//	 * @Auth(P_PUBLIC)
-//	 */
-//	public function login() {
-//		$form = new LoginForm(); // fetches POST values itself
-//		$values = $form->getValues();
-//
-//		if ($form->validate() && $this->loginService->login($values['user'], $values['pass'])) {
-//			if ($values['remember']) {
-//				$remember = $this->rememberLoginRepository->nieuw();
-//				$this->rememberLoginRepository->rememberLogin($remember);
-//			}
-//
-//			if ($values['redirect']) {
-//				return $this->csrRedirect(urldecode($values['redirect']));
-//			}
-//			return $this->redirectToRoute('default');
-//		} else {
-//			if ($values['redirect']) {
-//				return $this->redirectToRoute('csrdelft_login_loginform', ['redirect' => $values['redirect']]);
-//			}
-//
-//			return $this->redirectToRoute('csrdelft_login_loginform');
-//		}
-//	}
+	/**
+	 * @Route("/login_check", name="app_login")
+	 * @Auth(P_PUBLIC)
+	 */
+	public function login(AuthenticationUtils $authenticationUtils): Response {
+		if ($this->getUser()) {
+			return $this->redirectToRoute('default');
+		}
 
-//	/**
-//	 * @return RedirectResponse
-//	 * @throws ORMException
-//	 * @throws OptimisticLockException
-//	 * @Route("/logout", methods={"GET","POST"})
-//	 * @Auth(P_LOGGED_IN)
-//	 */
-//	public function logout() {
-//		$this->loginService->logout();
-//		return $this->redirectToRoute('default');
-//	}
+		// get the login error if there is one
+		$error = $authenticationUtils->getLastAuthenticationError();
+		// last username entered by the user
+		$lastUsername = $authenticationUtils->getLastUsername();
+
+		// TODO doe hier iets mee
+
+		return $this->redirectToRoute('default');
+	}
+
+	/**
+	 * @Route("/login_check", name="app_login_check")
+	 * @Auth(P_PUBLIC)
+	 */
+	public function login_check() {
+		throw new \LogicException('Wordt opgevangen door de firewall.');
+	}
+
+	/**
+	 * @Route("/logout", name="app_logout")
+	 * @Auth(P_PUBLIC)
+	 */
+	public function logout() {
+		throw new \LogicException('This method can be blank - it will be intercepted by the logout key on your firewall.');
+	}
 }
