+++ lib/view/login/LoginForm.php
@@ -2,15 +2,15 @@
 
 namespace CsrDelft\view\login;
 
-use CsrDelft\common\ContainerFacade;
-use CsrDelft\service\security\LoginService;
 use CsrDelft\view\formulier\CsrfField;
 use CsrDelft\view\formulier\elementen\HtmlComment;
 use CsrDelft\view\formulier\Formulier;
+use CsrDelft\view\formulier\invoervelden\HiddenField;
 use CsrDelft\view\formulier\invoervelden\TextField;
 use CsrDelft\view\formulier\invoervelden\WachtwoordField;
 use CsrDelft\view\formulier\keuzevelden\CheckboxField;
 use CsrDelft\view\formulier\knoppen\LoginFormKnoppen;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Security\Http\Authenticator\FormLoginAuthenticator;
 
 /**
@@ -20,14 +20,22 @@ use Symfony\Component\Security\Http\Authenticator\FormLoginAuthenticator;
  */
 class LoginForm extends Formulier {
 
-	public function __construct($showMelding = false) {
-		parent::__construct(null, '/login_check');
+	/**
+	 * @var UrlGeneratorInterface
+	 */
+	private $urlGenerator;
+
+	public function __construct(UrlGeneratorInterface $urlGenerator, $csrfToken, $error, $showMelding = false) {
+		parent::__construct(null, $urlGenerator->generate('app_login_check'));
 		$this->formId = 'loginform';
 		$this->showMelding = $showMelding;
 
 		$fields = [];
 
-		$fields[] = new CsrfField(ContainerFacade::getContainer()->get('security.csrf.token_manager')->getToken('authenticate'), '_csrf_token');
+		$redirectUri = filter_input(INPUT_GET, 'redirect', FILTER_UNSAFE_RAW);
+		$fields['redirect'] = new HiddenField('redirect', $redirectUri);
+
+		$fields[] = new CsrfField($csrfToken, '_csrf_token');
 
 		$fields['user'] = new TextField('_username', null, null);
 		$fields['user']->placeholder = 'Lidnummer of emailadres';
@@ -35,8 +43,8 @@ class LoginForm extends Formulier {
 		$fields['pass'] = new WachtwoordField('_password', null, null);
 		$fields['pass']->placeholder = 'Wachtwoord';
 
-		if (ContainerFacade::getContainer()->get(LoginService::class)->hasError()) {
-			$fields[] = new HtmlComment('<p class="error">' . ContainerFacade::getContainer()->get(LoginService::class)->getError() . '</p>');
+		if ($error) {
+			$fields[] = new HtmlComment('<p class="error">' . $error . '</p>');
 		} else {
 			$fields[] = new HtmlComment('<div class="float-left">');
 			$fields[] = new HtmlComment('</div>');
@@ -45,6 +53,7 @@ class LoginForm extends Formulier {
 		}
 
 		$this->addFields($fields);
+		$this->urlGenerator = $urlGenerator;
 
 		$this->formKnoppen = new LoginFormKnoppen();
 	}
